Fixed problems related to parsing only the "section" within a Wikipedia talk page.
1.wikicode sometimes interpret sections as not "sections" but subparts of another section in cases when
parsing a whole page. To prevent this, grab the section first and parse that section only, by using "rvsection" in mediawiki query.
**Wikum's table 'website_article' needs a column 'section_index' for this.**

2.Removed a previously added function for getting rid of templates in order to prevent the whole RfC section
being interpreted as one comment.
In minor cases this blocks wikicode from parsing properly.
Therefore removed the function, but need to find another way to solve the problem of RfCs being interpreted as one blob of comment.

3. Added a helper function that uses recursion to find section within the given wikicode.
This is needed when RfCs exist as a subsection of a section.
However, need to write code to differentiate cases when the outer section needs to be included along with the RfC.

4. Removed the error thrown when multiple headings exist in a section from section.py in wikichatter.
It's okay in Wikum because import_data.py makes sure only the right section is parsed,
and it is inevitable in some RfCs to include multiple headings.